+++ Fluentver/Pages/Storage.xaml.cs
@@ -1,4 +1,4 @@
-namespace Fluentver.Views
+namespace Fluentver.Pages
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
