+++ Fluentver/Pages/Users.xaml.cs
@@ -1,6 +1,6 @@
 using Windows.System;
 
-namespace Fluentver
+namespace Fluentver.Pages
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
